Move ApiProperties to config package
