+++ src/main/kotlin/de/bund/digitalservice/useid/config/ApiProperties.kt
@@ -1,4 +1,4 @@
-package de.bund.digitalservice.useid.apikeys
+package de.bund.digitalservice.useid.config
 
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.stereotype.Component
